Add modal state management
Created a new file, 'modalStore.js' in the 'stores' folder. This file will manage the state of the modal which is whether the modal is open or not. Modified 'AppHeader.vue' to toggle the modal state on login/register link click, and 'Auth.vue' to control modal visibility based on the modal state. This was done to have a global state control the modal instead of handling it in each component.