"Refactored Auth component to improve modal functionality"
The changes made to the Auth component have introduced a transition effect and improved the way the modals are closed. Previously, the modals were governed by a computed property 'hiddenClass', which has been replaced with methods using a Modal Store. A closeModal method was also added to allow the closing of the modal by clicking on the overlay or the close button. This provides a better user experience as it makes it easier to close modals. The transition effect added gives the modal a smoother and more engaging opening and closing sequence.