Add functionality to toggle between login and register forms
The component Auth.vue was updated to allow for the user interface to toggle between a display of 'login' and 'register' forms depending on the selected tab. This was achieved by adding a new data property 'tab' and altering classes and visibility based on its value. This results in a more engaging and responsive user experience.