Implement separate login and registration components
Added two new Vue components LoginForm and RegisterForm. The reason behind this is to separate the concerns of both forms, previously both forms were handled in the single Auth.vue component. Additionally, to provide better form validation functionality, the validation configuration was modified in validation.js to validate fields on change and on blur. Also, the CSS for the submit button was moved from the Auth.vue component to the App.vue.

These modifications will result