Updated authentication functionality for user login/logout
The changes encompass updates to the loginForm and registerForm components in the Vue.js application, including the implementations of user login and logout functionalities. Specifically, the changes have been made to enhance the user’s experience during account creation, authentication, and session termination.

The logic of the login and register forms was extended by adding userStore functionality which increases application maintainability. This will help the application vertically manage user’s sessions. A store was specifically imported into the login form for account authentication and into register form for account creation.

Furthermore, the AppHeader component was updated to display different links based on whether or not a user is authenticated. If