Update CompositionItem and Manage components
Refactored CompositionItem and Manage components to add delete functionality and prevent navigation with unsaved changes. The CompositionItem component was simplified and its methods were updated for better response handling. The Manage component was updated to add route guards. This is for enhanced user interaction and making sure the user does not lose unsaved changes by accident. Further, a delete function was added to enable the deletion of songs.