Add infinite scrolling to song list
Relocated the fetching of song data to a method called getSongs so it can be reused. Implemented scroll event listener to handle infinite scrolling of songs. On reaching the bottom of page, a console message is logged, this can be replaced in the future with a call to fetch more songs. This change improves user experience by providing smoother navigation in song list.