Add functionality to sort songs
Added a new feature that allows the user to sort songs by providing an additional query parameter, `sort` in the URL. This parameter accepts two values, 1 and 2. If any other value or no value is provided, it defaults to 1. This enhances UX by giving users control over the display order of songs.