Add audio playing functionality using Howl.js
A new file, `player.js`, has been added to manage the audio playing functionality, specifically to handle the action of a new song. It uses the Howler.js library to play the audio. This file is now imported into the `Song.vue` file, and a play functionality is added to the play button. Furthermore, Howler.js version 2.2.3 is added to the project dependencies in `package.json` file.