Add player progress bar functionality and improve UI
Added a feature to the player component where a progress bar updates as the song plays. The progress bar reflects the current time of the song and updates the seek and duration indicators. Introduced the new helper function formatTime to transform the seek and duration time from seconds into a more readable format. Also added another class to the play button for a clearer UI. This functionality will improve the UX as it gives users a visual indicator of a song's progress and allows them to seek the audio to a specified time.