Add direct link to comments section
Adjusted the Song and SongItem components to create a direct link to the comments section of a song. The 'comments' section in Song.vue is now assigned an id, which is used in SongItem.vue to create a link that takes the user directly to the comments section. It improves the user experience by making navigation within the song details page more seamless.