Add dynamic font-awesome icon directive
A new Vue directive 'icon' has been added to handle dynamic font-awesome icons. This change simplifies the way icons are being handled in the codebase, making it more reusable and maintainable. With this, we can use `v-icon="'icon-name'"` to add a Font Awesome icon, moving the logic from inline HTML to Vue directive. Updated HomeView and main.js to apply this new directive.