Improve icon directive with additional binding modifiers
The changes allow for the addition of new modifiers for the positioning and color of the icon. The updated 'icon.js' file now has right and yellow as binding modifiers. If the right modifier is present, the icon gets the additional class 'float-right' which moves it to the right side of its container. The color is set to green by default but is changed to yellow if the yellow modifier is present. The applied changes make the icon implementation more flexible and consistent across various elements of the app, in this case, as seen in the 'HomeView.vue' file.