Add secondary icon directive
A new directive, 'icon-secondary', was added. This directive allows to decorate specific elements with secondary icons. The directive provides a configurable option to place the icon on the right side of the element. This change was made to make the UI more interactive and visually pleasing. The directive is used in the 'HomeView' component for the 'Songs' list title.