Add i18n support and translation files for English and Dutch
This commit adds internationalization support to the app by introducing vue-i18n as a dependency and creating the corresponding translation files for English and Dutch (en.json, nl.json). The application's primary language is set to English, but it now supports switching to the Dutch language dynamically. The i18n instance is created in the `i18n.js` file within the 'includes' directory and is integrated into the Vue app in the 'main.js' file. The 'HomeView.vue' has been updated to use the translated text. This change will allow the application to support multilingual user interfaces and be more accessible to a wider audience.