Add internationalization and currency adaptation
This commit includes the addition of keypath for i18n translation for the Terms of Service in the registration form. This allows the ability to translate the text into other languages when necessary. It also encloses the Terms of Service text with a link tag.

In addition, setup for multi-currency adaptation was added to the song info, showing a price in the desired currency format, ultimately increasing user-friendliness for international users. I18n numberFormats for currencies in USD and JPY were added as a starting point.