Add localization for "Manage" and "About" sections
Added localization keywords for "Manage" and "About" under "header" and "manage" sections in both `en.json` and `nl.json` files. This change allows for present string literal to be converted to a localized version based on the chosen language setting. These adjustments are applied in `Manage.vue` and `AppHeader.vue` files. The purpose is to provide a more user-friendly navigation experience for the users of different language preferences.