Add NProgress package to project for enhanced loading UX
The package `nprogress` was added to the project (found in `package-lock.json` & `package.json`) to improve the user experience by displaying a progress bar during page loads. The `nprogress.css` was imported in `src/main.js` to provide necessary styling for the progress bar and the functionality to start and stop the progress bar during router transitions was implemented in a new file `src/includes/progress-bar.js`.